Fixes units of jsd_read timeout in Process()
Sets the EtherCAT cycle period to be the value of the timeout when fetching the EtherCAT frame via the jsd_read function. The underlying SOEM's function, `ecx_receive_processdata` expects the timeout to be in microseconds. The issue was that `Manager::Process` passed a timeout expressed in nanoseconds. Hence, the fix was to express such quantity in microseconds.