Actuator optional limits on actual position
Summary:
Adds optional parameters to specify limits on the actuator's actual
position. If the limits are violated, an all-device fault is issued
and motion is interrupted as a result.
The actuator can be driven back to within limits through a profile
position command with a target position that is within limits.
These limits are useful as a safeguard for non-position commands
which did not do checks on position, and as a replacement of
Elmo-level limits on position feedback which get disabled if the
modulo functionality is used (applicable to certain absolute
encoders).

Test Plan:
Tested on local network with a motor. Commanded the motor outside
the limits and verified the motor is stopped. Verified the motor
could be returned to within limits with a profile position command.

Reviewers: abrinkman, JosephBowkett